+++ src/ignitus-Authentication/ignitus-Common-Auth-Student-Prof/Components/constants.ts
